Support multiple consecutive linear solves
Initial values for a linear solve are now set in a `prepare` action
instead of in the initializer. This allows re-setting the linear solver
to its initial state to perform multiple consecutive solves.
This split also makes the order of initialization actions more
flexible, and fixes a bug where the initial residual was not
written to disk because it was observed before the registration phase.